pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ee-review"
version = "0.1.0"
description = "Full-stack web app with FastAPI and Vue"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pydantic",
    "passlib[bcrypt]",
    "python-jose[cryptography]"
]

[project.scripts]
ee_review = "backend.cli:main"

[tool.setuptools]
packages = ["backend"]
include-package-data = true

[tool.setuptools.package-data]
backend = ["static/**/*", "auth/*"]
